(PC-17103) auth: move useLoginRoutine to a separate file
