+++ src/features/auth/__mocks__/AuthContext.tsx
@@ -10,7 +10,3 @@ export const useAuthContext = jest.fn().mockReturnValue({
 export const signOut = jest.fn()
 
 export const useLogoutRoutine = jest.fn(() => signOut)
-
-export const loginRoutine = jest.fn()
-
-export const useLoginRoutine = jest.fn(() => loginRoutine)
