+++ src/features/auth/login/__mocks__/useLoginRoutine.ts
@@ -1,0 +1,1 @@
+export const useLoginRoutine = jest.fn(() => jest.fn())
